move publishing code into separate file
